+++ stories/02-Design System/01-General/Typography/Template.tsx
@@ -1,8 +1,8 @@
 import {ReactNode} from 'react';
 import {KitTypography} from '@kit/General/';
-import {KitGrid, KitSpace} from '@kit/Layout/';
 import {IEditorTemplate} from '../../../types';
 import {IKitTypography} from '@theme/types/general/typography';
+import {KitSpace} from '@kit/Layout';
 
 const getComponent = (component, content, args) => {
     switch (component) {
@@ -29,21 +29,44 @@ export const Template = args => {
 };
 
 export const EditorTemplate: IEditorTemplate = () => (
-    <KitGrid.Row>
-        <KitGrid.Col span={5}>
+    <div style={{width: '100%', display: 'grid', gridTemplateColumns: '50% 50%'}}>
+        <div>
             <KitSpace direction="vertical">
-                <KitTypography.Title>h1. Aristid Design</KitTypography.Title>
+                <KitTypography.Title>h1. Aristid Designs</KitTypography.Title>
                 <KitTypography.Title level="h2">h2. Aristid Design</KitTypography.Title>
                 <KitTypography.Title level="h3">h3. Aristid Design</KitTypography.Title>
                 <KitTypography.Title level="h4">h4. Aristid Design</KitTypography.Title>
             </KitSpace>
-        </KitGrid.Col>
-        <KitGrid.Col span={5}>
+        </div>
+        <div>
             <KitSpace direction="vertical">
                 <KitSpace direction="vertical" size="none">
                     <KitTypography.Text weight="bold">Aristid Design Text - bold</KitTypography.Text>
                     <KitTypography.Text weight="medium">Aristid Design Text - medium</KitTypography.Text>
                     <KitTypography.Text weight="regular">Aristid Design Text - regular</KitTypography.Text>
+                </KitSpace>
+                <KitSpace direction="vertical" size="none">
+                    <KitTypography.Text weight="regular" size="fontSize1">
+                        Aristid Design Text - fontSize 1
+                    </KitTypography.Text>
+                    <KitTypography.Text weight="regular" size="fontSize2">
+                        Aristid Design Text - fontSize 2
+                    </KitTypography.Text>
+                    <KitTypography.Text weight="regular" size="fontSize3">
+                        Aristid Design Text - fontSize 3
+                    </KitTypography.Text>
+                    <KitTypography.Text weight="regular" size="fontSize4">
+                        Aristid Design Text - fontSize 4
+                    </KitTypography.Text>
+                    <KitTypography.Text weight="regular" size="fontSize5">
+                        Aristid Design Text - fontSize 5
+                    </KitTypography.Text>
+                    <KitTypography.Text weight="regular" size="fontSize6">
+                        Aristid Design Text - fontSize 6
+                    </KitTypography.Text>
+                    <KitTypography.Text weight="regular" size="fontSize7">
+                        Aristid Design Text - fontSize 7
+                    </KitTypography.Text>
                 </KitSpace>
                 <KitSpace direction="vertical" size="none">
                     <KitTypography.Paragraph weight="bold">Aristid Design Paragraph - bold</KitTypography.Paragraph>
@@ -58,37 +81,8 @@ export const EditorTemplate: IEditorTemplate = () => (
                     <KitTypography.Link weight="regular">Aristid Design Link - regular</KitTypography.Link>
                 </KitSpace>
             </KitSpace>
-        </KitGrid.Col>
-        <KitGrid.Col span={8}>
-            <KitSpace direction="vertical" size="none">
-                <KitTypography.Text size="fontSize1">Aristid Design Text - fontSize1</KitTypography.Text>
-                <KitTypography.Text size="fontSize2">Aristid Design Text - fontSize2</KitTypography.Text>
-                <KitTypography.Text size="fontSize3">Aristid Design Text - fontSize3</KitTypography.Text>
-                <KitTypography.Text size="fontSize4">Aristid Design Text - fontSize4</KitTypography.Text>
-                <KitTypography.Text size="fontSize5">Aristid Design Text - fontSize5</KitTypography.Text>
-                <KitTypography.Text size="fontSize6">Aristid Design Text - fontSize6</KitTypography.Text>
-                <KitTypography.Text size="fontSize7">Aristid Design Text - fontSize7</KitTypography.Text>
-            </KitSpace>
-            <KitSpace direction="vertical" size="none">
-                <KitTypography.Paragraph size="fontSize1">Aristid Design Paragraph - fontSize1</KitTypography.Paragraph>
-                <KitTypography.Paragraph size="fontSize2">Aristid Design Paragraph - fontSize2</KitTypography.Paragraph>
-                <KitTypography.Paragraph size="fontSize3">Aristid Design Paragraph - fontSize3</KitTypography.Paragraph>
-                <KitTypography.Paragraph size="fontSize4">Aristid Design Paragraph - fontSize4</KitTypography.Paragraph>
-                <KitTypography.Paragraph size="fontSize5">Aristid Design Paragraph - fontSize5</KitTypography.Paragraph>
-                <KitTypography.Paragraph size="fontSize6">Aristid Design Paragraph - fontSize6</KitTypography.Paragraph>
-                <KitTypography.Paragraph size="fontSize7">Aristid Design Paragraph - fontSize7</KitTypography.Paragraph>
-            </KitSpace>
-            <KitSpace direction="vertical" size="none">
-                <KitTypography.Link size="fontSize1">Aristid Design Link - fontSize1</KitTypography.Link>
-                <KitTypography.Link size="fontSize2">Aristid Design Link - fontSize2</KitTypography.Link>
-                <KitTypography.Link size="fontSize3">Aristid Design Link - fontSize3</KitTypography.Link>
-                <KitTypography.Link size="fontSize4">Aristid Design Link - fontSize4</KitTypography.Link>
-                <KitTypography.Link size="fontSize5">Aristid Design Link - fontSize5</KitTypography.Link>
-                <KitTypography.Link size="fontSize6">Aristid Design Link - fontSize6</KitTypography.Link>
-                <KitTypography.Link size="fontSize7">Aristid Design Link - fontSize7</KitTypography.Link>
-            </KitSpace>
-        </KitGrid.Col>
-    </KitGrid.Row>
+        </div>
+    </div>
 );
 EditorTemplate.path = 'components.Typography';
 EditorTemplate.title = 'Typography';