+++ stories/02-Design System/02-Layout/Grid/Grid.stories.tsx
@@ -2,6 +2,7 @@ import type {Meta, StoryObj} from '@storybook/react';
 import {KitGrid} from '@kit/Layout';
 import {argTypes} from './data';
 import {Template, FakeGrid} from './Template';
+import GridTest from './test-components/gridTest';
 
 const meta: Meta<typeof FakeGrid> = {
     component: FakeGrid,
@@ -17,4 +18,8 @@ export const Api: Story = {
     parameters: {
         chromatic: {disableSnapshot: true}
     }
+};
+
+export const ChromaticTest: Story = {
+    render: () => <GridTest />
 };