+++ stories/02-Design System/04-DataEntry/Checkbox/examples/index.tsx
@@ -4,7 +4,6 @@ import CheckAll from './checkAll';
 import Disabled from './disabled';
 import Danger from './danger';
 import Group from './group';
-import Grid from './grid';
 import LabelHelper from './labelHelper';
 
 import BasicSource from './basic?raw';
@@ -13,7 +12,6 @@ import CheckAllSource from './checkAll?raw';
 import DisabledSource from './disabled?raw';
 import DangerSource from './danger?raw';
 import GroupSource from './group?raw';
-import GridSource from './grid?raw';
 import LabelHelperSource from './labelHelper?raw';
 
 export const Sources = {
@@ -23,7 +21,6 @@ export const Sources = {
     Disabled: DisabledSource,
     Danger: DangerSource,
     Group: GroupSource,
-    Grid: GridSource,
     LabelHelper: LabelHelperSource
 };
 export default {
@@ -33,6 +30,5 @@ export default {
     Disabled,
     Danger,
     Group,
-    Grid,
     LabelHelper
 };